Attempt to resolve sys/stat related build issue on Alpha.
Not having access to an Alpha makes this difficult, but one thing
that is clearly not helping is that with Linux glibc on Alpha, we
end up defining *both* HAVE_STAT_TIMESPEC and HAVE_STAT_TIME_T, &
that breaks otherwise valid pmns.c coding assumptions.

This change means only one type for the stat time fields can be
defined, and it errs on the side of timespec (which should be the
one Alpha uses, AFAICT) - so perhaps this issue will finally be
resolved and PCP can progress in the Debian build/release system.