Fix the Windows SQL Server buffer_cache metric for more platforms.
This metric is a (32 bit float) PERF_RAW_FRACTION metric under
Windows 2003 Server and SQL Server 2000, but under Windows and
SQL Server 2008 its a 64 bit double (PERF_LARGE_RAW_FRACTION).
Needed additional code to handle this type, which we've not come
across before.

In addition reworked startup slightly to reduce the warnings and
correct types being used in the metric table for several Windows
metrics - when going to 64 bit Windows 2008, lots of types have
changed and pmcd.log was spammed with warnings as a result.  This
is now figured out on the fly, at PMDA startup, for the affected
metrics (using M_AUTO64 flag to identify those that will change).