ReplacePmnsSubtree fix bad signal handling botch
The trap handling was totally wrong in cleanup() ... causing the
command "1" to be executed instead of the null command "".

Exposed by another bug in a related area ... if the PMNS lock
exists already, after the timeout in pmlock we fall into the
trap cleanup() routine instead of charing on, so the script
fails horribly (and leaves the stale lock file around).

Also fixed an unlikely buffer over-run in mmv.c.