libpcp_pmda, dbpmda Minor additions for dynamic subtrees in the PMNS
In libpcp_pmda the wrong error was being returned if an attempt was
made to translate a dynamic metric name to a PMID, and the PMDA is NOT
using PMDA_INTERFACE_4 (very low chance of this happening).

And the additional documentation for dbpmda to describe the new commands
to exercise the services a PMDA needs to provide to play in the dynamic
subtrees of the PMNS (this was just forgotten in the "final" checkin
for these changes).