Fix pmcd.timezone value so it updates with daylight savings.
The root of the problem is that __pmTimezone writes the value
of TZ that it returns into the environment as well, causing
subsequent callers to always find the value of the first call
instead of querying the system for the current timezone.

Fix is to clear TZ setting before calling __pmTimezone, which
demonstrably fixes the problem.  Note some platforms (Windows)
do not have unsetenv() so a suitable fallback is used in that
situation.