Don't change pointer types of pmCtime's arguments
On 64 bit Linux and Solaris platforms time_t is 64 bit integer, __pmTimeval
assumes that time is 32 bit integer, the code was too liberal in passing
pointers to the later into function which was expecting former. On at least
one occasion this was causing segmentation violations in pmCtime when
localtime(3) refused to deal with rubbish in the upper bytes of the 64 bit
integer.