getpmcdhosts etc rework -v pcp=2 support
Deprecate -v ipc=... and -v libpcp=... as we're not using these and there
is no easy way to determine these versions.

For -v pcp the =1 or =2 support has been changed to be <op><number>,
e.g. -v 'pcp>=2'.

All of this needed for PCP 3.0 where many of the QA tests need to be
run on hosts with PCP 2.x _or_ PCP 3.x