pmlogextract change timezone selection algorithm, add -f for old behaviour
When more than one timezone is present in the input archives, we need
to choose one timezone for the output archive.  The new default is to
use the timezone from the last input archive, with the -f option being
added to revert to the previous behaviour of choosing the timezone from
the first input archive.

Since pmlogextract is the only tool to create <mark> records, added verbage
in the man page to explain what this is and why this is done.