pmlc sometimes connnects to wrong pmcd
This was causing QA/381 to fail sometimes (depending on the setup
of the remote qa hosts).

After connecting to pmlogger, pmlc may need to connect to the same
pmcd that pmlogger is connected to.  When pmlogger is connected to
pmcd on "localhost" and pmlc is run from a different machine, then
pmlc may connect to the wrong pmcd or no pmcd at all, depending on
the state of the host where pmlc is run.