pmdaproc.sh Fix pminfo usage
pminfo is used in several places within the script ...
	pminfo
	pminfo -h localhost
	pminfo -n $NAMESPACE
	pminfo -h localhost -n $NAMESPACE

Remove all the -h localhost stuff as this is simply not needed.

Remove the -n $NAMESPACE stuff as this is wrong (need to be interrogating
pmcd's view of the PMNS) and this does not work correctly in the presence
of metrics that use the dynamic PMNS features.

Also cleaned up some of the "Usage" reporting.