Convert pmdammv to using the new dynamic namespace interfaces.
This change convert pmdammv to use the libpcp_pmda dynamic namespace
interfaces instead of the traditional pmcd namespace file updating
mechanism.   For a PMDA like MMV, the old way results in spurious
errors (EAGAIN) whenever a metadata change occurs, due to the nature
of the exchange with pmcd resulting in the namespace being updated.

This is inappropriate for a production PMDA, and is now resolved.
As another benefit of these changes, we are now able to provide a
shared library variant of pmdammv for all platforms - this patch
also includes that change.