pmdaproc.sh Cleanup
Significant change is to _not_ use -n option to pminfo unless -n was
specified in the command line.  This is needed to avoid warnings for
PMDAs that install dynamic metrics into the PMNS (like the new mmv).

Also dropped -h localhost (this is the default).

And added -V option to allow more verbose output (at this stage the
full output from pminfo after a PMDA install with metrics that have
errors or warnings).