Add PMLOGGER_REQUEST_TIMEOUT support
Setting $PMLOGGER_REQUEST_TIMEOUT in  the environment defines a number
of seconds for the timeout to be applied to interactions between pmlc
and pmlogger ... specifically, pmlc will timeout in a __pmGetPDU call
if the pmlogger is not responding.

Added __pmLoggerTimeout() to libpcp as checks need to be made in libpcp
(logconnect.c and logcontrol.c) and in pmlc (mostly in actions.c)

Requested by Nathan after strange failure of pmlogger monitoring
scripts in their production environment when pmlc connected to a
morbid pmlogger but the pmlc hung trying to get information from
pmlogger.