mmv.reload help text changes
Because mmv.reload is below the root of a dynamic PMNS subtree at
mmv, the name mmv.reload cannot be resolved by newhelp before the
PMDA is installed.

So the PMDA needs to supply the help text for _all_ metrics of the
form mmv.* ... and the "help" file is not needed.