Make pmdacisco work for our local office router (2960 model).
There's a problem in the parsing code where both a username and password
are required.  The username extraction ends with the Password: prompt, &
consumes this token - the password extraction starts by looking for that
token - so it hangs indefinately in the interface probe stage.  Addressed
this by passing a pointer to the "Password:" string, if it was observed,
so that the next stage can (optionally) start with that token.  Otherwise
everything is done exactly as before.

Cleansed explicit hostnames from the pmdacisco Samples/Tested files also.