Improve pcp-glider commands interaction with Cygwin shell.
When running pcp command shell scripts from a Cygwin/ssh shell, we
also want to use the sh that comes with Glider, not the Cygwin sh.
The .bat wrapper trick to make this work under cmd.exe doesn't fly
for Cygwin shell, as it picks the shell scripts up ahead of Windows
batch files (ugh).  We work around this by adding in a SHELLSUFFIX
macro along the lines of EXECSUFFIX, and on windows install with a
.sh suffix.  That way the .bat files are found first again, and all
is well once more.  Probably other Cygwin interaction issues will
crop up down the track, but this gets us started.