Integrate PCP into Solaris Service Management Facilicy
Add postinstall/preremove scripts to update SMF entries for pmcd, pmie
and pmproxy. On install import entires into SMF but do not automatically
enable services - leave it to sysadmin to do so. On remove disable services
and remove entires.

Also propogate install and default targets into SUBDIRS of build: I
needed to install a file which logically should live in the build/sun
and I did not fancy to do the install by hand.