Support configure --libexecdir for $PCP_BINADM_DIR
Adds support for PCP_BINADM_DIR=/usr/libexec/pcp/bin on platforms that
use libexec, and use it explicitly in the Fedora RPM spec, where we
now also make use the %{_libexecdir} RPM macro.

This whole section of configure.in could do with some simplification
and rationalization, but that's not the intent of this commit.

	modified:   CHANGELOG
	modified:   build/rpm/pcp_fedora.spec
	modified:   configure.in
	modified:   src/include/pcp.conf.in

Signed-off-by: Mark Goodwin <mgoodwin@redhat.com>