mmv Install issue fixed
Since the mmv PMDA was promoted to one of the "install by default"
PMDAs, the Install script comman no longer worked ... a fall out from
going from root + pmns to describe the PMDA's PMNS to using root_mmv.

The fix is to extend pmnsadd to recognize when it is given a complete
PMNS to work with, rather than the usual subtree of the PMNS case and
make sure root_mmv is available in the PMDA directory.