readline for dbpmda
Thanks to Andrew Wansink fo showing how this could be done and providing
the initial patches.  This checkin refines that work so that readline()
is used when it is available (and when the lexer is flex) but only when
the command input is coming from a tty device.

This version also works on vanilla Mac OS X.

The same approach could also be used also for the non-flex lexers,
but I don't have any way of testing these.