Fix View Save handling of instance names with embedded square braces.
The original code truncates instance names from the initial closing
brace, which means attempts to subsequently use that view fail.  We
now move backward from the end of the metric[instance] string until
we find the final closing brace (at the end of the string).

Fixes a problem reported with Windows network interfaces, thanks to
Nigel Donaldson for reporting it.