Differentiate between PM_ERR_INST and PM_ERR_PMID in pmdammv.
The return code from the fetch callback was being unilaterally
set to PM_ERR_PMID in pmdammv, if its just an instance missing
we can now tell (and thats less of an issue).  Parfait writes
such mmv(4) files, sometimes; checking with Cowan whether that
is expected behaviour there or not (not ideal obviously but it
may be unavoidable).