Generalized DSO PMDA support for PM_CONTEXT_LOCAL
This batch of changes adds to the existing hardcoded table (and funky
environment variable rules) that control which DSO PMDAs may be used
with PM_CONTEXT_LOCAL.

Specifically, __pmLocalPMDA() is added to libpcp to allow updates to the
table, i.e. add a new entry, delete an existing entry, clear all entries.

The monitoring tools that understand -L from the command line
(pminfo and pmprobe) or support the equivalent @:metric syntax from
pmParseMetricSpec() (pmval) have been given a new command line option,
-K to get access to __pmLocalPMDA().  pmstat groks -L, but the metrics
it uses are hardcoded to the OS PMDA, so there was no point in addding
-K for pmstat.

In the process the #define PM_USE_CONTEXT_LOCAL (which was always set)
was removed from the source allowing some dead code to be removed.

Some minor clean up of the man pages also was done, e.g. documenting
the -c option for pminfo (hang over from the derived metrics changes).