Change how PM_CONTEXT_LOCAL determines available PMDAs
Moving from the previous ugly build-time configuration of dsotbl[]
to run-time determination based on pmcd.conf.

To make a new PMDA available to applications using PM_CONTEXT_LOCAL
simply install the PMDA as a "dso" ... this updates the pmns and pmcd.conf
and the PMDA will be added to the list of PMDAs available with
PM_CONTEXT_LOCAL contexts.

In the process, rolled to PCP 3.1.3 to help isolate QA differences.