Fix local context reporting of help text.
Help text coming out of the direct PMDA calls to get help text
is not dynamically allocated, so we cannot free it.  pmLookupText
callers (like pminfo) are expecting the buffer to be allocated
however, so we need to strdup the text before passing it back.

MMV and Windows PMDAs need updating as they are leaking memory
in this area at the moment (will be separate commits).