Improve pmlogger handling of alias names (same PMID, different names)
Change the way pmlogger writes out metric names to ensure that
the requested metric name (in configuration file) is recorded
in the archive always.  This change also removes the protocol
exchange between pmlogger and pmcd on receipt of a first fetch
for a metric, since that information (names and descriptor) is
actually already cached and was unnecessary.