Fix a free of non-malloc'd memory in dbpmda metric help text code
Turns out pmcd is expecting the text pointers to not be allocated
dynamically (contrary to the man page), and if they are they will
be leaked (fixed earlier).  But dbpmda was expecting these to be
alloc'd and was free'ing them - which causes dso pmdas to barf as
soon as text requested (free of mmap'd ptr, typically).

This makes dbpmda behave the same as pmcd, and updates the manual
page for pmdaText accordingly.

Fixed up a couple of hyphen-vs-minus-sign manual page issues too.