650 make it faster and working for others
Reduced the logging interval by a factor of 5 to speed up the test.

Replaced all the reported instances and values for pmcd.agent.status and
the alias sample.secret.foo.bar.max.redirect to be something more generic,
as the instances and values logged depend on the local environment
outside QA's control.