Export information about ZFS snapshots
Count number of snapshots in each ZFS filesystem and export information
about snapshots. The snapshots have their own instance domain (the
names are different from the filesystems) but are fecthed together with
the filesystems to which the snapshots have been applied. I've considered
creating separate methodtab entry for snapshots but decided agaist it
because it would require extra scan for all the filesystems and also because
it make it more convoluted when calculating the number of snapshots
per filesystem.

Also export extra metrics like refreservation and usedbyrefreservation
for the filesystems.