Allow use of SunStudio compiler on Solaris
I've been building Solaris packages using SunStudio compilers for
almost a year (had to do it because my perl was build using SunStudio
compilers and MakeMaker uses options of the compiler it was built with)
by overriding configure's selection via CC in the environment, now it's
time to make it more official: SunStudio installs suncc as symlink to
cc - by giving AC_PROC_CC the list of compilers to try I can use that
one in preference to GNU compiler. Because the PATH used by configure
is not propagated to the rest of the build use AC_PATH_PROG to provide
full path to the compiler location for the rest of the build via
builddefs.

Also check if host supports 64 bit and if it does then build
using 64 bit mode but allow user to override it by using --with-64bit
option to configure.