Make log metadata reading more robust in a specific corruption case.
We switch metric IDs behind the backs of some metric names recently.
This results in archives which have some metric names with two PMIDs
which isn't great but isn't totally fatal either - the archive can
still be used for all other metrics, and even the metric in question
is visible for part of the archive.  The current behaviour (which is
to refuse to read the log at all) is not ideal, so now we allow the
archive to still be loaded in this situation.