pmie - fix [no]match_inst botch, tweak sleepTight reporting
Problem with match_inst that was reported by Max with a metric having
only 1 instance was a day one botch (using the wrong expression node)
... not sure how this worked reliably in the past.  New QA 358 was able
to reproduce this problem reliably before the fix.

Also relaxed the reporting of negative delays in sleepTight() ... small
negative values can occur if the sampling frequency is very high and there
are multiple fetch groups, or in the presence of pmcd timeouts ... both
appear in the QA tests.