Add utility to extract data from pmie stats files
Greg Banks reported that root on his machine was spammed by pmie_check: his
machine only has 3 letters in its hostname and this is not long enough for
string(1) used by pmie_check to consider it a real string. Sure, I could've
used strings -n2 to reduce the limit but given that it is a second attempt
of solve the problem of parsing binary files with inadequate tools
I've decided to to create a proper tool - pmiestatus.