Complete the pmdalinux cgroup code for user-defined cgroups.
Now ready for a release.  Handles the tricky case of having
to extend and rewrite parts of the metric table for metrics
which suddenly appeared/disappeared that cannot be catered
for in the static table.

This also fixes the percpu cpu-accounting metrics - need to
ensure we refresh that indom whenever those metrics are in
the requested set.