Clean up pmns debris on upgrade
On Solaris upgrade is performed by removing the old version
and installing the new one which, unfortunately, clobbers any
changes pmcd.conf. At the same time it doesn't clobber the
changes to pmns and leaves punter with a very confusing
setup - metrics in the namespace but no pmda to serve them.