pmchart - vertical axis unit scaling
Improved algorithm to adjust the PCP units for the vertical axis
so the axis label and axis scale values will fit within the available
real estate.  Previously we could have values up to 10,000 and the
units lable of Kbytes/sec were chopped at the left margin ... this will
now be rescales to Mbytes/sec with a maximum value of 1,000 which
fits and reads better.