Merge pmchart and pmtime adp (qt assistant) files.
Now have a single PCP Manual pcpdoc.adp file which both
pmchart and pmtime (and any other new tools) will use.
Update both applications to point to this new file, and
fix up the pkg_icon_dir configure reference therein so
that it points to the right location.

Renamed importdata so that it is named consistently with
the other tutorials.