Make pmwtf report when one/other metric average changes from zero.
Since we currently only report ratios of one to another, the code
was simply skipping metrics that change to/from a zero average,
as that introduces divide-by-zero issues.  However there's critical
information there, needs to be reported, so added such a mechanism
which is easily distinguished from the ratio reporting.