Split Solaris' vnode operation counters into two instance domains
Every time a filesystem is mounted new device number is allocated
for it and that confused pmdaCache operations: even though
the name upto the first space was unique, the full instance name
wasn't the same and it was refusing to add that instance to the
cache.

Also gnb convinced me that mixing mountpoints and filesystem
types in one instance domain is lame, so I've separated the two
and now there is kernel.fs for counters for individual filesystems
and kernel.fstype for counters for each fileystem type.