Use devinfo library to get information about pretty disk names
Translating from disk names used by ZFS to the instances in disk instance
domain is not always straight forward: pretty disk names use bus address
to identify target and bus address is not always the same as instance in
kstat. Libdevinfo provides an API (a convoluted one) to get the path to
the symlink under /dev from the first minor device for a particular disk.