Add Qt Frameworks into Mac OS X package
Copy the Qt frameworks into the package to make it independent of
the Qt version - use can install pcp-gui without installing Qt or
with different version of Qt.

The frameworks are copied during the install phase into temporary
location, the path to dependencies is modified in both framework and
the binary and then the whole shebang is installed. I did consider
modifying install-sh to allow installation directly from the absolute
path but because the framework's dynamic libraries must be modified
using the system libraries is not the best of ideas.