Sweep through every QA test and make tests consistent in several areas.
All uses of sudo no longer assume sudo lives in the topdir, so we can
now switch to the using the sudo supplied with the system (README also
updated to describe trivial /etc/sudoers entry needed to make it work).
Since this touches almost every script, made the following changes while
in there:
- get $sudo from a single location (a common script), ensure all use it
  consistently
- every test also defines $tmp, and almost every test defined $here, so
  make these come from the same place too (tmp can be moved if needed)
- remove old rcs $Revision lines as these are no longer useful with git
- complete purging of owner line, was inconsistently used if at all, and
  no longer has any meaning.
- remove duplicate common file sourcing - in particular, localconfig is
  already included via common.product, tons of tests reimported it.

Finally, had to move the "broken" test pmda into a directory named with
pmdas/broken in the path, as the Install script common code now expects
this convention to be followed.