Resolve an issue in the rpm generation process, from man pages.
The configure'd man page path was sometimes evaluating with NONE
as a component, which ulitimately lead to rpm generation failure.

Reworked the man page format and location detection logic to use
the value pcp was configured with, rather than having pcp-gui try
to second guess it.  While here, added in support for bz2 manual
page compression which some distributions use.