libpcp, pmcd, dbpmda, pmlogger, et al - retire PDU_ASCII and PDU_BINARY
PDU_ASCII was effectively retired some time ago.  Vestiges of this
and the only supported PDU format (PDU_BINARY) remained in the source.
This commit cleans this up.

Since we only have binary PDUs, the "mode" argument to all of the
__pmDecode* routines is not needed ... remove it!

PM_ERR_NOASCII is not used, pmErrStr() message changed to reflect
the fact that this error code is no longer used.

In several places PDU_* was overloaded, fixed by introducing new
macros relevant to the specific use.  Changes as follows ...

__pmGetPDU
	Second arg is now ANY_SIZE (was PDU_BINARY) or LIMIT_SIZE
	(was PDU_CLIENT) ... the latter is used in pmcd and pmproxy
	to avoid DOS attacks with malcrafted PDUs.

dbpmda
	Use new macros NO_CONN (was PDU_NOT), CONN_DAEMON  (was
	PDU_BINARY) and CONN_DSO (was PDU_DSO) to indicate type and
	state of PMDA connection.

pmcd
	Don't need pduProtocol in AgentInfo structure.

pmlogger
	Left over from last round of changes, used FROM_ANON instead
	of PDU_BINARY for __pmSend* calls.