Change end-of-context protocol for daemon PMDAs
The previous implementation required the PMDA to send an ERROR PDU as an
ack after receiving the end-of-context notification from PMCD or dbpmda.
This was (a) unnecessary, and (b) meant PMCD was seeing a lot of pduread()
timeouts waiting for a response from a daemon PMDA that may not even
have been using PMDA_INTERFACE_5 or later (PMCD cannot tell the version
of libpcp_pmda, if any, the PMDA is using in this case).

Drop the ack PDU.