404 - extend coverage to error cases
pmUnpackEventRecords has a lot of internal checks for the consistency
of the packed array of event records ... exercise this logic.

Also the valgrind suppression fingerprint for the one (so far) known
false hit is slightly different for binaries built and installed from a
workarea versus those installed from a package (at least on Ubuntu 10.10).