pmdaproc.sh - perl Install fixups
The installation logic for a Perl PMDA was clobbering the pmns and
domain.h files ... this is sort of OK if the PMDA is _only_ ever
installed as a Perl PMDA, but for schizo cases like the simple PMDA
this cripples any attempt to install the PMDA as a non-Perl agent.

Just be more careful with the files being created, and also need to
reorder some large blocks of functionality in the install case for
all PMDAs.