__pmRegisterAnon() - event record rework
Redo the anonymous metrics implementation to allow different names
to be registered.

Change the format for packed event records so that "missed" records
are encoded as an event record with er_flags set to PM_ER_FLAG_MISSED
and er_nparams set to the number missed records.

Also changed the format for the unpacked pmResult delivered via
pmUnpackEventRecords() so that er_flags is presented as an additional
event parameter for the anonymous metric event.flags and when missed
records are encountered, the number of missed records is exported as
an additional event parameter for the anonymous metric event.missed.