PM_TYPE_AGGREGATE and __pmStuffValue changes
This is the mooted "Option 3".

Change the internal encoding of PM_TYPE_AGGREGATE metrics via
a pmAtomValue - void *vp replaced by pmValueBock *vbp.

Remove the unnecessary aggr_len parameter to __pmStuffValue
(was only useful for PM_TYPE_AGGREGATE metrics and now the length
is encoded in the vbp->vlen header of the value passed via a
pmAtomValue.

All the other modifications are consequential on these two changes.