Resolve issue with Windows service failing to stop pmcd.
This ultimately turned out to be related to shell scripts
not being able to find pmsignal.sh, thanks to aliases not
getting expanded in non-login shells.  Resolved by using
a shell function instead.

As I introduced a Windows-specific pcp.env to handle this,
took the opportunity to remove the PCP_CONFIG hack which
had started to spread like mould.  Now only PCP_CONF and
PCP_DIR need be set on Windows (installer does this auto-
magically, but still).