Add perly bits into Darwin's build manifest
Perl modules use their own Makefiles which are generated by MakeMaker.
Those makefiles do not follow PCP rules and do no use install-sh to
push the new files into the staging tree, which means that after running
make install the files are not included in the manifest and are not packaged.
Debian and rpm builds use custom rules to harvest those perly bits, now
Darwin joins them - on Darwin if DIST_MANIFEST is set then the extra
pass is done to insure that modules, their manpages and shared libraries are
included in the manifest.

I've considered creating separate package just for perl bits but
decided against it, at least now - the support infrastructure in Darwin/MacOS
packaging is too involved to allow for simple parameterization.