Consolidate to common Win32 lib and include directories.
All the Win32 code (including kernel PMDA) now uses the shared
libraries and headers.  The headers are updated to be copies of
those included with the latest mingw-w64 snapshot (once we get
the next Strawberry Perl update, these should all disappear as
everything will come from there then) instead of the homebrew
variants we had until now.

Added in the Event Tracing for Windows headers and TDH library,
as well, for initial experimentation.