Correct cross-platform herrno handling, particularly for Win32.
This commit adds the initial macros for correcting cross-platform
issues with how we deal with errno and friends.  It introduces a
set of platform independent macros which abstract the differences
and does the translation for herrno.  Various scattered hacks to
deal with hstrerror throughout the code have beem removed.

Changes for errno and the Win32 socket API error codes are alot
more extensive, thats not addressed here (that will come next).