libpcp - thread-safe and refactoring changes
Next steps towards a thread-safe libpcp - this time adding "const" to
a range of declarations.

Make identification of the library for pthreads consistent with the
other optional libraries in the build.

Introduce a "big lock" to be used as a global mutex in libpcp.

Refactor libpcp to remove some replicated code and move the functionality
into libpcp, see __pmPrintStamp() and __pmPrintTimeval().  Also move the
"hash" methods into their own source file.