Improve robustness of Win32 trace consumer exercising tool.
In particular, when it crashes it doesn't get a chance to switch
off the trace session(s) it had enabled.  In the case of the Win32
kernel event tracer, there can be only one, so subsequent attempts
to run the tool always failed.  This issue is now resolved and no
longer plagues us.