libpcp - multi-threaded support
Moving to a more general mechanism for ensuring single-threaded
usage in the places where we cannot or should not make the code truly
multi-threaded.

Allowing multiple "scopes" for single-threaded enforcement such that
only one thread is allowed to call the routines within a scope, but it
is possible for different threads to call into different scopes.