354 - Mac OS X tweaking
Unsetting $DISPLAY to cause pmconfirm to fail does not work on Mac OS
X, so use $PCP_XCONFIRM_PROG and a custom QA shell script to exit so
that pmlogger will notice and report the message that would normally be
part of a pmconfirm dialog.

Also, the fields of the ps(1) output are different for Mac OS X, so
handle this as well.