Several additional metrics for log file PCP agent.
Add in counts of bytes and events seen per logfile, and the
current log file size for each.  This will aid debugging and
day-to-day monitoring of more complex, hierarchical (r)syslog
deployments, for example.

Added a fetch callback so initial code to check status of the
monitored log files could be added - will need to extend this
though, to deal with the common case of logfile rotation.

Minor cleanups - pmid_string field is not a string, rename;
do not explicitly initialise global variables to zero, as the
compiler will happily do that; metric semantics for a couple
of metrics not-quite-right (discrete vs instant).