local context issues fixed
Serious problem leading to segv in libpcp when using __pmLocalPMDA() to
add/delete DSOs for PM_CONTEXT_LOCAL use ... when the context is closed
at application exit the call via e_endCallBack may have been made for
a dso table entry that was no longer valid.

And in a related area, the parsing, error detection and error reporting
in __pmSpecLocalPMDA() was greatly improved.