Resolve a couple of issues from missing log file handling code.
If we do not stat the file on initial opening as well, the first
ino/dev comparison fails and we end up always reopening the file
at least once.

Whenever we have a successful stat call on logfile, keep latest
output in case we ever want to use last-modified times, etc.