More improvements to the TDH exerciser.
Worryingly, fixes a bug in some sample code that could
never have worked - pinfo must be passed by reference
into GetEventInformation if we're to use it on return.

ProcessTrace can only deal with one realtime session,
so update code to reflect that.

Finally, use a well-known named session for all events
that will be consumed by PCP - will allow us to offload
the (very complex and provider specific task of) event
enabling to existing windows tools that already do that.