Add memory limiting functionality to pmdalogger.
Support for capping the amount of memory that will be held
for events, preventing a slow-client-fast-event-generator
situation from causing pmdalogger memory consumption issues.

Added in two new metrics to help track this - global amount
of maximum memory for event data (per file) and the amount
used in each event queue.