Extend the event store mechanism to be more generally useful.
Instead of just storing an integer 1 to the event trace metric,
send an arbitrary string.  Contents and interpretation are still
PMDA-specific, but this allows much more interesting filtering
of events on the server side (instead of just being a basic way
to restrict access).

We could extend pmdalogger to filter based on a regex this way,
later on, once it becomes more line-oriented.  More interesting
uses may exist for future PMDAs - this could provide a way for
systemtap/dtrace scripts to be sent to the server from clients,
for example.