Fix RUN_IN_BUILD_ENV for Win32 builds with pmcpp.
The ways of Windows path resolution are subtle and bizarre
in the extreme.  Using "../" (dot,dot,forward-slash) is not
allowed apparently (backslashes are OK with relative paths)
and the DOS shell (cmd.exe) produces this error:

"'..' is not recognized as an internal or external command,
operable program or batch file."

We can workaround this by using a path from the drive root,
even though it still contains "../" components.