__pmInitLocks - change initialization
Previously we were using run-time initialization to ensure
__pmLock_libpcp is a recursive mutex.

As of Ubuntu 11.04 and/or NPTL 2.13 as reported by
    $ getconf GNU_LIBPTHREAD_VERSION
this was not working, and the mutex was being created as
a normal mutex, leading to deadlock with out inevitable nested
lock() calls form the one thread in libpcp.

So changed to compile-time intialization using
    PTHREAD_RECURSIVE_MUTEX_INITIALIZER
from <pthreads.h> if that macro is defined, else fallback to
run-time initialization.