libpcp - static buffer changes
All of these routines return a string value in a static buffer and hence
are not thread-safe: pmAtomStr, pmEventFlagsStr, pmIDStr, pmInDomStr,
pmNumberStr, pmTypeStr and pmUnitsStr.

This commit adds Berkeley-style equivalents for each so
	pmFooStr(args)
is accompanied by the thread-safe version
	pmFooStr_r(args ..., char *buf, int buflen)
to allow the value to be returned in a buffer managed by the caller.

The same change was made for the (undocumented) internal routines
__pmPDUTypeStr, __pmLogName and __pmTimezone which also return values
via a static buffer.

Also change all the internal calls to these routines in libpcp and
the other PCP libraries to use the thread-safe variants.

A related change was to the static method statestr() in optfetch.c
where the function was changed to use a caller-provided buffer
rather than a local static buffer.

In libpcp/src/tz.c there are even more extensive changes to use
the big library lock to make references to the static variables
and the calls to unsafe libc routines all thread-safe.