Use direct event pointer in pmdalogger last event tracking.
Instead of using pointer-to-a-pointer, easier to use a direct
pointer.  Makes it a bit simpler to follow the code and the
TAILQ macros from queue.h can be used (more).

Also added a diagnostic routine such that we only print the
first few characters of event parameter string, to make the
agent log file easier to follow.