Rework Kens deb packaging fix for config files.
lintian didn't like the previous way, and turns out theres
a simpler way - just list the extra conf files in a special
debian/pcp.conffiles and the existing scripts do the right
thing.

Added lintian overrides for this and other non-issues, but
left the remaining "real" ones still to be worked through.