Implement logger PMDA line-oriented event mode.
Change pmdalogger to generate individual events records
for each new line in the log file, instead of the current
read(2)-sized chunks.  This paves the way for application
of per-client regular expressions for event filtering.

This is loosely based on the similar code in the Perl PMDA
wrapper module.

Also here, turns out the pipe inode mtime is not updated
on every write ... so we only skip (non-blocking) reads on
inodes we know to be regular files, else we miss updates.