making tarball installations work better
Add helper scripts for tarball-based packaging.

Package upgrades should proceed as:
	# here=`pwd`
	# ./preinstall
	# ( cd /; tar xpf $here/pcp-*.tar.gz )
	# ./postinstall

Package removal can be done from the same directory using
	# ./remove

The helper scripts understand about the various init/rc frameworks
as well as the PCP configuration files that may have been changed and
need to be reinstated across an upgrade.

The remove is driven of the manifest of files in the tarball.